api/requirements/test.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements/test.in
#
aiocontextvars==0.2.2     # via databases
alembic==1.1.0
asn1crypto==0.24.0        # via cryptography
asyncpg==0.18.3           # via databases
atomicwrites==1.3.0       # via pytest
attrs==19.1.0             # via packaging, pytest
bcrypt==3.1.7             # via paramiko, passlib
cached-property==1.5.1    # via docker-compose
certifi==2019.6.16        # via requests
cffi==1.12.3              # via bcrypt, cryptography, pynacl
chardet==3.0.4            # via requests
click==7.0                # via pip-tools, uvicorn
contextvars==2.4          # via aiocontextvars
cryptography==2.7         # via paramiko
databases[postgresql]==0.2.5
dataclasses==0.6          # via pydantic
dnspython==1.16.0         # via email-validator
docker-compose==1.24.1    # via pytest-docker-compose
docker-pycreds==0.4.0     # via docker
docker[ssh]==3.7.3        # via docker-compose
dockerpty==0.4.1          # via docker-compose
docopt==0.6.2             # via docker-compose
email-validator==1.0.4
fastapi==0.36.0
h11==0.8.1                # via uvicorn
httptools==0.0.13         # via uvicorn
idna==2.7                 # via email-validator, requests
immutables==0.9           # via contextvars
importlib-metadata==0.19  # via pluggy, pytest
jsonschema==2.6.0         # via docker-compose
mako==1.1.0               # via alembic
markupsafe==1.1.1         # via mako
miracle-acl==0.0.4.post1
more-itertools==7.2.0     # via pytest, zipp
packaging==19.1           # via pytest
paramiko==2.6.0           # via docker
passlib[bcrypt]==1.7.1
pip-tools==4.1.0
pluggy==0.12.0            # via pytest
psycopg2-binary==2.8.3    # via databases
psycopg2==2.8.3
py==1.8.0                 # via pytest
pycparser==2.19           # via cffi
pydantic==0.30            # via fastapi
pyjwt==1.7.1
pynacl==1.3.0             # via paramiko
pyparsing==2.4.2          # via packaging
pytest-docker-compose==3.1.0
pytest==5.1.1
python-dateutil==2.8.0    # via alembic
python-editor==1.0.4      # via alembic
python-multipart==0.0.5
pyyaml==3.13
requests==2.20.1
six==1.12.0               # via bcrypt, cryptography, docker, docker-compose, docker-pycreds, dockerpty, packaging, pip-tools, pynacl, python-dateutil, python-multipart, websocket-client
sqlalchemy==1.3.8
starlette==0.12.7         # via fastapi
texttable==0.9.1          # via docker-compose
urllib3==1.24.3           # via requests
uvicorn==0.8.6
uvloop==0.12.2            # via uvicorn
wcwidth==0.1.7            # via pytest
websocket-client==0.56.0  # via docker, docker-compose
websockets==7.0           # via uvicorn
zipp==0.6.0               # via importlib-metadata
